feat: add order detail components and utilities
- Introduced new components for order details: OrderHeader, OrderProgress, OrderStatusCards, OrderTimeline, OrderShipment, OrderItems, and OrderSummary.
- Added utility functions for formatting currency and dates, and determining status colors.
- Updated OrdersList to remove unnecessary whitespace.
- Enhanced useCustomerOrders hook for better readability and structure.
- Created order.schema.ts for order-related schemas.
