+++ apps/web/src/components/orders/detail/index.ts
@@ -1,3 +1,11 @@
 export { default as OrderOverview } from "./OrderOverview";
 export { default as PaymentUpload } from "./PaymentUpload";
 export { default as ConfirmButton } from "./ConfirmButton";
+export { default as OrderHeader } from "./OrderHeader";
+export { default as OrderProgress } from "./OrderProgress";
+export { default as OrderStatusCards } from "./OrderStatusCards";
+export { default as OrderTimeline } from "./OrderTimeline";
+export { default as OrderShipment } from "./OrderShipment";
+export { default as OrderItems } from "./OrderItems";
+export { default as OrderSummary } from "./OrderSummary";
+export * from "./orderUtils";
