+++ apps/web/src/hooks/useCustomerOrders.ts
@@ -77,7 +77,7 @@ async function fetchOrders(
   } = {}
 ): Promise<OrdersResponse> {
   const params = new URLSearchParams();
-  
+
   if (typeof filters.page === "number")
     params.append("page", String(filters.page));
   if (typeof filters.pageSize === "number")
@@ -89,7 +89,6 @@ async function fetchOrders(
 
   const currentUserId = getCurrentUserId();
   const url = `http://localhost:8000/api/orders?${params.toString()}`;
-  
 
   const response = await fetch(url, {
     headers: {
@@ -103,7 +102,6 @@ async function fetchOrders(
   }
 
   const data = await response.json();
-  
 
   // API returns: { message: "...", data: { items, total, page, pageSize, pagination } }
   if (data.data) {
@@ -113,19 +111,32 @@ async function fetchOrders(
   return { items: [], total: 0, page: 1, pageSize: filters.pageSize ?? 10 };
 }
 
-export function useCustomerOrders({ page, pageSize, status, q, dateRange }: UseCustomerOrdersParams) {
+export function useCustomerOrders({
+  page,
+  pageSize,
+  status,
+  q,
+  dateRange,
+}: UseCustomerOrdersParams) {
   return useQuery({
-    queryKey: ["customer-orders", page, status, q, dateRange.from, dateRange.to],
+    queryKey: [
+      "customer-orders",
+      page,
+      status,
+      q,
+      dateRange.from,
+      dateRange.to,
+    ],
     queryFn: async () => {
       let dateFrom: string | undefined;
       let dateTo: string | undefined;
-      
+
       if (dateRange.from) {
         const startOfDay = new Date(dateRange.from);
         startOfDay.setHours(0, 0, 0, 0);
         dateFrom = startOfDay.toISOString();
       }
-      
+
       if (dateRange.to) {
         const endOfDay = new Date(dateRange.to);
         endOfDay.setHours(23, 59, 59, 999);
@@ -135,11 +146,11 @@ export function useCustomerOrders({ page, pageSize, status, q, dateRange }: UseC
         endOfDay.setHours(23, 59, 59, 999);
         dateTo = endOfDay.toISOString();
       }
-      
+
       if (dateFrom || dateTo) {
         // Date range filter is applied; keep behavior but don't log in production
       }
-      
+
       return fetchOrders({
         page,
         pageSize,
